rename IoU -> Jaccard Index (#662)
* rename IoU -> Jaccard Index
* deprecate
* imports
