+++ torchmetrics/__init__.py
@@ -14,7 +14,7 @@ _PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
 from torchmetrics import functional  # noqa: E402
 from torchmetrics.aggregation import CatMetric, MaxMetric, MeanMetric, MinMetric, SumMetric  # noqa: E402
 from torchmetrics.audio import PIT, SDR, SI_SDR, SI_SNR, SNR  # noqa: E402
-from torchmetrics.classification import (  # noqa: E402
+from torchmetrics.classification import (  # noqa: E402, F401
     AUC,
     AUROC,
     F1,
@@ -31,6 +31,7 @@ from torchmetrics.classification import (  # noqa: E402
     HammingDistance,
     Hinge,
     IoU,
+    JaccardIndex,
     KLDivergence,
     MatthewsCorrcoef,
     Precision,
@@ -101,7 +102,7 @@ __all__ = [
     "FBeta",
     "HammingDistance",
     "Hinge",
-    "IoU",
+    "JaccardIndex",
     "KLDivergence",
     "MatthewsCorrcoef",
     "MaxMetric",
